+++ lib/core/data.py
@@ -93,6 +93,7 @@ options: dict[str, Any] = {
     "redirects_history": False,
     "color": True,
     "quiet": False,
+    "disable_cli": False,
     "output_file": None,
     "output_format": None,
     "log_file": None,