+++ lib/core/options.py
@@ -36,7 +36,7 @@ from lib.parse.nmap import parse_nmap
 
 
 def parse_options() -> dict[str, Any]:
-    opt = parse_config(parse_arguments())
+    opt = merge_config(parse_arguments())
 
     if opt.session_file:
         if opt.async_mode:
@@ -200,12 +200,32 @@ def parse_options() -> dict[str, Any]:
         )
         exit(1)
 
-    if opt.output_format not in OUTPUT_FORMATS:
-        print(
-            "Select one of the following output formats: "
-            f"{', '.join(OUTPUT_FORMATS)}"
-        )
+    opt.output_formats = [format.strip() for format in opt.output_formats.split(",")]
+    invalid_formats = set(opt.output_formats).difference(OUTPUT_FORMATS)
+
+    if invalid_formats:
+        print(f"Invalid output format(s): {', '.join(invalid_formats)}")
         exit(1)
+
+    # There are multiple file-based output formats but no variable to separate output files for different formats
+    if (
+        opt.output_file
+        and "{format}" not in opt.output_file
+        and len(opt.output_formats) - ("mysql" in opt.output_formats) - ("postgresql" in opt.output_formats) > 1
+        and (
+            "{extension}" not in opt.output_file
+            # "plain" and "simple" have the same file extension (txt)
+            or {"plain", "simple"}.issubset(opt.output_formats)
+        )
+    ):
+        print("Found at least 2 output formats sharing the same output file, make sure you use '{format}' and '{extension} variables in your output file")
+        exit(1)
+
+    if opt.log_file:
+        opt.log_file = FileUtils.get_abs_path(opt.log_file)
+
+    if opt.output_file:
+        opt.output_file = FileUtils.get_abs_path(opt.output_file)
 
     return vars(opt)
 
@@ -247,12 +267,13 @@ def _access_file(path: str) -> File:
         return fd
 
 
-def parse_config(opt: Values) -> Values:
+def merge_config(opt: Values) -> Values:
     config = ConfigParser()
     config.read(opt.config)
 
     # General
     opt.thread_count = opt.thread_count or config.safe_getint("general", "threads", 25)
+    opt.async_mode = opt.async_mode or config.safe_getboolean("general", "async")
     opt.include_status_codes = opt.include_status_codes or config.safe_get(
         "general", "include-status"
     )
@@ -319,8 +340,8 @@ def parse_config(opt: Values) -> Values:
     opt.suffixes = opt.suffixes or config.safe_get("dictionary", "suffixes", "")
     opt.lowercase = opt.lowercase or config.safe_getboolean("dictionary", "lowercase")
     opt.uppercase = opt.uppercase or config.safe_getboolean("dictionary", "uppercase")
-    opt.capitalization = opt.capitalization or config.safe_getboolean(
-        "dictionary", "capitalization"
+    opt.capital = opt.capital or config.safe_getboolean(
+        "dictionary", "capital"
     )
 
     # Request
@@ -357,23 +378,25 @@ def parse_config(opt: Values) -> Values:
 
     # Advanced
     opt.crawl = opt.crawl or config.safe_getboolean("advanced", "crawl")
-    opt.async_mode = opt.async_mode or config.safe_getboolean("advanced", "async")
 
     # View
     opt.full_url = opt.full_url or config.safe_getboolean("view", "full-url")
     opt.color = opt.color if opt.color is False else config.safe_getboolean("view", "color", True)
     opt.quiet = opt.quiet or config.safe_getboolean("view", "quiet-mode")
+    opt.disable_cli = opt.disable_cli or config.safe_getboolean("view", "disable-cli")
     opt.redirects_history = opt.redirects_history or config.safe_getboolean(
         "view", "show-redirects-history"
     )
 
     # Output
-    opt.output_path = config.safe_get("output", "autosave-report-folder")
-    opt.autosave_report = config.safe_getboolean("output", "autosave-report")
-    opt.log_file_size = config.safe_getint("output", "log-file-size")
-    opt.log_file = opt.log_file or config.safe_get("output", "log-file")
-    opt.output_format = opt.output_format or config.safe_get(
-        "output", "report-format", "plain", OUTPUT_FORMATS
+    opt.output_file = opt.output_file or config.safe_get("output", "output-file")
+    opt.mysql_url = opt.mysql_url or config.safe_get("output", "mysql-url")
+    opt.postgres_url = opt.postgres_url or config.safe_get("output", "postgres-url")
+    opt.output_table = config.safe_get("output", "output-sql-table")
+    opt.output_formats = opt.output_formats or config.safe_get(
+        "output", "output-format", "plain"
     )
+    opt.log_file = opt.log_file or config.safe_get("output", "log-file")
+    opt.log_file_size = config.safe_getint("output", "log-file-size")
 
     return opt